refactor(visitas): consolida o cast no modulo de acesso (remove builder duplicado do hook)
